front: refacto Scenario component - create a new ScenarioContent component which is rendered only if a scenario, a timetable and an infra exist - modify useScenarioData so scenario, timetable and infra are arguments of the hook
